refactor v2 lookup eval tests
